Add disable sync step to gmail
On some devices the sync step does not complete and causes the workload
to error. Add a step to disable syncing on gmail by unchecking the sync
option in the app settings.

Note: An initial sync is still required when the app is first started
but for subsequent iterations syncing will be turned off.