WA3 Hello AR workload
This workload runs custom variations of the ARCore sample app Hello_AR. These apps should have been compiled with specific settings for ARCore. This workload selects the correct APK based on the provided parameters from the name of the APK which should comply to the format `com.google.ar.core.examples.java.helloar_XYZ` where `XYZ` are the capitalized initials of the parameters. The workload launches the app, waits for a provided duration before closing the app. After running the app, the workload extracts from the logcat numeric values sitting at the end of lines containing the keyword `UXPERF` and preceded by either `FrameTime` or `PointCloudTime` and, depending on the `export_csv_times` parameter, outputs these in csv files respectively named `frametimes.csv` and `cloudtimes.csv`. Finally it adds the averages of these signals and their corresponding frequencies to the iteration metrics.